Updated all floats to Doubles (#257)
* Updated tests

Changed all floats into doubles to have consistency over the whole framework

Made a small update to the multiplexer to better push through supply and demand

Fixed small typo

Updated M3SA paths.

fixed merge conflicts

Removed unused components. Updated tests.

Improved checkpointing model

Improved model, started with SimPowerSource

implemented FailureModels and Checkpointing

First working version

midway commit

first update

All simulation are now run with a single CPU and single MemoryUnit. multi CPUs are combined into one. This is for performance and explainability.

* Updated test memory

+++ opendc-compute/opendc-compute-simulator/src/main/java/org/opendc/compute/simulator/telemetry/GuestCpuStats.java
@@ -38,6 +38,6 @@ public record GuestCpuStats(
         long idleTime,
         long stealTime,
         long lostTime,
-        float capacity,
-        float usage,
-        float utilization) {}
+        double capacity,
+        double usage,
+        double utilization) {}
