+++ opendc-compute/opendc-compute-simulator/src/main/kotlin/org/opendc/compute/simulator/telemetry/table/HostInfo.kt
@@ -30,6 +30,6 @@ public data class HostInfo(
     val name: String,
     val arch: String,
     val coreCount: Int,
-    val coreSpeed: Float,
+    val coreSpeed: Double,
     val memCapacity: Long,
 )
